Find OpenMP manually for microcontrollers Only find and link std::filesystem when not compiling for microcontrollers Minor fixes

+++ include/hipSYCL/sycl/context.hpp
@@ -146,7 +146,7 @@ public:
                 << std::endl;
           }
           
-          last_platform = rt::platform_id{b, platform_index};
+          last_platform = rt::platform_id{b, static_cast<int>(platform_index)};
           found_device_platform = true;
         }
       }
